Year 2019 Day 22 Part 1

+++ src/main/java/net/eugenpaul/adventofcode/helper/HashStorage.java
@@ -31,7 +31,7 @@ public class HashStorage {
         return false;
     }
 
-    public String getHashOf(int step) {
+    public String getHashOf(long step) {
         if (cycleBegin == null) {
             return storage.entrySet().stream().filter(v -> v.getValue() == step).findFirst().orElseThrow().getKey();
         }
